[refactor] create MatchSettingCommonLayout and refactor CreateMatchScreen
