+++ feature/src/main/java/com/begin_a_gain/feature/match/match/MatchScreen.kt
@@ -82,7 +82,7 @@ fun MatchScreen() {
         showBackButton = true,
         trailingIcon = OImageRes.Menu,
         onTrailingIconClick = {
-            // Todo : navigate to menu screen
+
         },
         useDefaultPadding = false,
         snackBarBottomPadding = 104.dp
